test(chaos): consolidate fs chaos and modular automation test suites
